get all resorts in App.js
